LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-inside-outside-expected.txt


FAIL .target 1 assert_equals:
<div class="target" style="left:anchor(inside)" data-offset-x="150"></div>
offsetLeft expected 150 but got 0
FAIL .target 2 assert_equals:
<div class="target" style="left:anchor(outside)" data-offset-x="200"></div>
offsetLeft expected 200 but got 0
FAIL .target 3 assert_equals:
<div class="target" style="right:anchor(inside)" data-offset-x="190"></div>
offsetLeft expected 190 but got 390
FAIL .target 4 assert_equals:
<div class="target" style="right:anchor(outside)" data-offset-x="140"></div>
offsetLeft expected 140 but got 390
FAIL .target 5 assert_equals:
<div class="target" style="top:anchor(inside)" data-offset-y="250"></div>
offsetTop expected 250 but got 0
FAIL .target 6 assert_equals:
<div class="target" style="top:anchor(outside)" data-offset-y="300"></div>
offsetTop expected 300 but got 0
FAIL .target 7 assert_equals:
<div class="target" style="bottom:anchor(inside)" data-offset-y="290"></div>
offsetTop expected 290 but got 390
FAIL .target 8 assert_equals:
<div class="target" style="bottom:anchor(outside)" data-offset-y="240"></div>
offsetTop expected 240 but got 390
FAIL .target 9 assert_equals:
<div class="target" style="inset-inline-start:anchor(inside)" data-offset-x="150"></div>
offsetLeft expected 150 but got 0
FAIL .target 10 assert_equals:
<div class="target" style="inset-inline-start:anchor(outside)" data-offset-x="200"></div>
offsetLeft expected 200 but got 0
FAIL .target 11 assert_equals:
<div class="target" style="inset-inline-end:anchor(inside)" data-offset-x="190"></div>
offsetLeft expected 190 but got 390
FAIL .target 12 assert_equals:
<div class="target" style="inset-inline-end:anchor(outside)" data-offset-x="140"></div>
offsetLeft expected 140 but got 390
FAIL .target 13 assert_equals:
<div class="target" style="inset-block-start:anchor(inside)" data-offset-y="250"></div>
offsetTop expected 250 but got 0
FAIL .target 14 assert_equals:
<div class="target" style="inset-block-start:anchor(outside)" data-offset-y="300"></div>
offsetTop expected 300 but got 0
FAIL .target 15 assert_equals:
<div class="target" style="inset-block-end:anchor(inside)" data-offset-y="290"></div>
offsetTop expected 290 but got 390
FAIL .target 16 assert_equals:
<div class="target" style="inset-block-end:anchor(outside)" data-offset-y="240"></div>
offsetTop expected 240 but got 390
